requirements.txt
tornado~=6.1
tortoise-orm
aiomysql
pyyaml
cryptography